requirements.txt
langchain    #library or framework
langchain-core    #package
langchain-community
PyPDF2
numpy
pandas
openai
langchain-openai
#PyMuPDF
#faiss-cpu
chromadb
python-dotenv
fastapi
uvicorn
sentence-transformers
langchain_groq